Renamed COPYING to COPYING.template.
The repository does not enforce a specific license. Each patch may have their
own license/conditions depending on the contributors / copyright holders.

However, if a patch should ever be included in the official driver, it is
required to be licensed under the 2-clause BSD license as given in
COPYING.template. Therefore we strongly encourage contributors to use this
license for all patches contributed here.